pyproject.toml
[project]
name = "schoolcafebot"
version = "0.1.0"
description = ""
authors = [
    {name = "midry"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "discord-py (>=2.5.2,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
